+++ src/com/intellij/idea/plugin/hybris/impex/lang/annotation/ImpexAnnotator.kt
@@ -25,6 +25,7 @@ import com.intellij.idea.plugin.hybris.impex.psi.*
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpExHeaderAbbreviationReference
 import com.intellij.idea.plugin.hybris.lang.annotation.AbstractAnnotator
 import com.intellij.lang.annotation.AnnotationHolder
+import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.elementType
@@ -49,6 +50,18 @@ class ImpexAnnotator : AbstractAnnotator(DefaultImpexSyntaxHighlighter.instance)
 
     override fun annotate(element: PsiElement, holder: AnnotationHolder) {
         when (element.elementType) {
+            ImpexTypes.VALUE_LINE -> {
+                val valueLine = element as? ImpexValueLine ?: return
+                val headerLine = valueLine.headerLine ?: return
+                val valueLineIndex = headerLine.valueLines.indexOf(valueLine)
+
+                if ((valueLineIndex + 1) % 2 == 0) {
+                    highlight(ImpexHighlighterColors.VALUE_LINE_EVEN, holder, element, HighlightSeverity.INFORMATION)
+                } else {
+                    highlight(ImpexHighlighterColors.VALUE_LINE_ODD, holder, element, HighlightSeverity.INFORMATION)
+                }
+            }
+
             ImpexTypes.USER_RIGHTS_HEADER_PARAMETER -> {
                 val headerParameter = element as? ImpexUserRightsHeaderParameter ?: return
                 val elementType = headerParameter.firstChild.elementType ?: return